Make "size" members of type usize
The C API exposes blazesym_sym_info type with a size member of type u64.
We use usize for various sizes so let's do the same here. I went ahead
and converted a few more internal types as well. I preserved the
blazesym_sym_info.file_offset as u64 because that's how Rust generally
represents file offset (e.g.,
https://doc.rust-lang.org/std/io/enum.SeekFrom.html).

+++ src/elf/parser.rs
@@ -479,7 +479,7 @@ impl ElfParser {
                         found.push(SymbolInfo {
                             name: name.to_string(),
                             address: sym_ref.st_value as Addr,
-                            size: sym_ref.st_size,
+                            size: sym_ref.st_size as usize,
                             sym_type: SymbolType::Function,
                             file_offset: 0,
                             obj_file_name: None,
@@ -526,7 +526,7 @@ impl ElfParser {
                     syms.push(SymbolInfo {
                         name: sname.to_string(),
                         address: sym_ref.st_value as Addr,
-                        size: sym_ref.st_size,
+                        size: sym_ref.st_size as usize,
                         sym_type: SymbolType::Function,
                         file_offset: 0,
                         obj_file_name: None,
